+++ bridge/l2/watcher.go
@@ -43,7 +43,7 @@ type WatcherClient struct {
 	messengerABI     *abi.ABI
 
 	// The height of the block that the watcher has retrieved event logs
-	processedMsgHeight uint64
+	processedMsgHeight int64
 
 	stopped uint64
 	stopCh  chan struct{}
@@ -65,7 +65,7 @@ func NewL2WatcherClient(ctx context.Context, client *ethclient.Client, vp *viper
 		ctx:                ctx,
 		Client:             client,
 		orm:                orm,
-		processedMsgHeight: uint64(savedHeight),
+		processedMsgHeight: savedHeight,
 		vp:                 vp,
 		messengerAddress:   messengerAddress,
 		messengerABI:       bridge_abi.L2MessengerMetaABI,
@@ -179,7 +179,7 @@ func (w *WatcherClient) fetchContractEvent(blockHeight uint64) error {
 		log.Info("l2 watcher fetchContractEvent", "w.processedMsgHeight", w.processedMsgHeight)
 	}()
 
-	fromBlock := int64(w.processedMsgHeight) + 1
+	fromBlock := w.processedMsgHeight + 1
 	toBlock := int64(blockHeight)
 
 	contractEventsBlocksFetchLimit := w.vp.GetInt64("contract_events_blocks_fetch_limit")
@@ -210,7 +210,7 @@ func (w *WatcherClient) fetchContractEvent(blockHeight uint64) error {
 			return err
 		}
 		if len(logs) == 0 {
-			w.processedMsgHeight = uint64(to)
+			w.processedMsgHeight = to
 			continue
 		}
 		log.Info("received new L2 messages", "fromBlock", from, "toBlock", to, "cnt", len(logs))
@@ -241,7 +241,7 @@ func (w *WatcherClient) fetchContractEvent(blockHeight uint64) error {
 			return err
 		}
 
-		w.processedMsgHeight = uint64(to)
+		w.processedMsgHeight = to
 	}
 
 	return nil
